+++ src/ImpromptuModular.hpp
@@ -173,7 +173,7 @@ inline bool calcWarningFlash(long count, long countInit) {
 NVGcolor prepareDisplay(NVGcontext *vg, Rect *box, int fontSize);
 
 void printNoteNoOct(int note, char* text, bool sharp);
-void printNote(float cvVal, char* text, bool sharp);
+int printNote(float cvVal, char* text, bool sharp);
 
 int moveIndex(int index, int indexNext, int numSteps);