+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.6 (in development)
 
 - Fixed bug in Foundry CUST selection (custom length selection for SEL button and copy-paste)
+- Add +/- suffix in display for split point when showing notes
 
 
 ### 1.1.5 (2019-03-25)